pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-optimizer"
version = "0.1.0"
description = "A CLI tool that optimizes prompts for AI assistants using Azure OpenAI"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "rich>=13.0",
    "questionary>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
prompt-optimizer = "prompt_optimizer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
